pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gs-convert"
version = "0.1.0"
description = "Convert modern images to Apple IIgs Super High-Resolution format"
authors = [{name = "Christopher Brown"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["apple2", "apple-iigs", "retro", "image-conversion", "dithering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
ui = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[project.scripts]
gs-convert = "gs_convert.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/gs_convert"
Documentation = "https://github.com/yourusername/gs_convert#readme"
Repository = "https://github.com/yourusername/gs_convert"
Issues = "https://github.com/yourusername/gs_convert/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
